rename spack flake8 to spack style
Currently aliases flake8 to style, and just runs flake8 as before.  Will
add mypy support after this.
